Worm 및 관련 부속 데이터, 애니메이션, 어빌리티 추가.

AttributeSet 확장: Damage, Exp 등 새로운 어트리뷰트 정의 및 복제 처리 반영.

AttackTask, Attribute 관련 로직 개선 및 코드 정리.

기존 Enemy 데이터 구조(DataTable 포함) 및 속성 처리 방식 수정.

PoolSubsystem 및 SpawnSubsystem 내 로직 업데이트.

AEnemyBase를 APawn에서 ACharacter로 변경.

SkeletalMesh 및 Capsule 관련 변수 제거, 기본 ACharacter 구성 요소로 대체.

Burrow 관련 Capsule 접근 함수 수정 및 충돌 처리 방식 업데이트.

다양한 연관 Blueprint 및 DataTable 데이터 갱신.

기존 데이터테이블 대신 어트리뷰트 메타데이터를 사용합니다

+++ Source/TenTenTown/Enemy/Data/EnemyData.h
@@ -29,13 +29,13 @@ struct FEnemyData : public FTableRowBase
 	
 	//Attribute Set에 적용
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int32 MaxHP = 100;
+	float MaxHP = 100.f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int32 CurrentHP = 100;
+	float CurrentHP = 100.f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int32 Attack = 10;
+	float Attack = 10.f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	float MovementSpeed = 200;
+	float MovementSpeed = 200.f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	float AttackSpeed = 1.0f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
@@ -43,7 +43,7 @@ struct FEnemyData : public FTableRowBase
 	
 	//처치 시 골드,경험치 보상
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int32 Gold = 10;
+	float Gold = 10;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	int32 EXP = 10;
